Added Project views tests; test files reorg
